Add noscript fallback
This change adds a fallback "Learn the truth" href button if javascript isn't available.
The button links to "/the-truth", which redirects to the youtube video. (Doing this masks the actual link the user will end up in)
the-truth.html can be removed if the redirect is set up properly in the webserver.